engine: use uint8_t for type of pheromone

+++ engine/pheromone.cpp
@@ -3,7 +3,8 @@
 
 const GameObjectCategory Pheromone::s_category{"PHEROMONE"};
 
-Pheromone::Pheromone(double latitude, double longitude, Team &team, int type)
+Pheromone::Pheromone(double latitude, double longitude, Team &team,
+					 uint8_t type)
 	: GameObject{Pheromone::category(), latitude, longitude, 0}, m_life{300},
 	  m_team{team}, m_type{type} {
 	m_teamBase = &team;
